kreader: add BitmapHolder as reference couting hodler of bitmap
Currently, we always copy bitmap when we get/put cache, copy backend rendered
bitmap to frontend viewport, which is inefficient and unnecessary.
To avoid extra cost of bitmap copy, we add BitmapHolder as a reference counting holder of
bitmap, so we can pass bitmap by BitmapHolder object instead of copy.